change language auto-generate section to subsection.
